feat: rename ModalWindow folder
